Add generated C files of select dirs in Linux v5.5-rc2
These C files have been generated with:

https://github.com/jajajasalu2/cocci-type-test-suite

with a few minor modifications to generate logs of each
type picked up and printed.

The logs have been modified to remove some of the spatch/make output
(commands of the Makefile, which files have been processed,
the amount of time taken for processing, the generated patch).

Pretty printing errors in the C files have been marked with a
"// err" at the end of the declaration. To view all of these
errors, simply run a grep command as follows:

	grep -rnie "// err"

which will give the line numbers of the errors. Note that these
errors have been determined and marked manually, so they may not be
absolutely correct.

The C files have been generated by running the test suite
generation scripts against the following directories of the
Linux kernel v5.5-rc2:

	init/
	kernel/
	mm/

Signed-off-by: Jaskaran Singh <jaskaransingh7654321@gmail.com>